requirements.txt
Flask==3.1.2
SQLAlchemy==2.0.44
python-dotenv==1.0.0